+++ docs/make.jl
@@ -4,17 +4,15 @@ makedocs(
     modules = [Dubins],
     format = Documenter.HTML(
         mathengine = Documenter.MathJax(),
-        prettyurls = get(ENV, "CI", nothing) == "true"
+        prettyurls = get(ENV, "CI", nothing) == "true",
     ),
     sitename = "Dubins",
     authors = "Kaarthik Sundar",
     pages = [
         "Home" => "index.md",
         "API Documentation" => "api.md",
-        "Library" => "library.md"
-    ]
+        "Library" => "library.md",
+    ],
 )
 
-deploydocs(
-    repo = "github.com/kaarthiksundar/Dubins.jl.git"
-)
+deploydocs(repo = "github.com/kaarthiksundar/Dubins.jl.git")