+++ src/Dubins.jl
@@ -1,21 +1,17 @@
 module Dubins
 
-using Memento
+using StaticArrays
+import Logging
+import LoggingExtras
 
-const LOGGER = getlogger(@__MODULE__)
-Memento.setlevel!(LOGGER, "info")
-
-__init__() = Memento.register(LOGGER)
-
-"Suppresses information and warning messages output by Dubins, for fine grained control use the Memento package"
-function silence()
-    Memento.info(LOGGER, "Suppressing information and warning messages for the rest of this session.  Use the Memento package for more fine-grained control of logging.")
-    Memento.setlevel!(Memento.getlogger(Dubins), "error")
-end
-
-"alows the user to set the logging level without the need to add Memento"
-function logger_config!(level)
-    Memento.config!(Memento.getlogger("Dubins"), level)
+# Setup Logging
+include("logging.jl")
+function __init__()
+    global _DEFAULT_LOGGER = Logging.current_logger()
+    global _LOGGER = Logging.ConsoleLogger(;
+        meta_formatter = Dubins._dubins_metafmt,
+    )
+    return Logging.global_logger(_LOGGER)
 end