+++ src/path_fcns.jl
@@ -1,10 +1,13 @@
 
-export
-    dubins_shortest_path, dubins_path,
-    dubins_path_length, dubins_segment_length,
+export dubins_shortest_path,
+    dubins_path,
+    dubins_path_length,
+    dubins_segment_length,
     dubins_segment_length_normalized,
-    dubins_path_type, dubins_path_sample,
-    dubins_path_sample_many, dubins_path_endpoint,
+    dubins_path_type,
+    dubins_path_sample,
+    dubins_path_sample_many,
+    dubins_path_endpoint,
     dubins_extract_subpath
 
 """
@@ -17,43 +20,40 @@ A configuration is given by ``[x, y, \\theta]``, where ``\\theta`` is in radians
 * ``\\rho``      - turning radius of the vehicle
 * return    - tuple (error code, dubins path). If error code != 0, then `nothing` is returned as the second argument
 """
-function dubins_shortest_path(q0::Vector{Float64}, q1::Vector{Float64}, ρ::Float64)
+function dubins_shortest_path(q0::VF, q1::VF, ρ::F, loop_tol::F = 1e-5) where {F,VF<:AbstractVector{F}}
 
     # input checking
-    @assert length(q0) ==  3
+    @assert length(q0) == 3
     @assert length(q1) == 3
-    (ρ <= 0) && (return EDUBBADRHO, nothing)
+    (ρ <= 0) && (return EDUBBADRHO, DubinsPath{F}())
 
-    path = DubinsPath()
-
-    params = zeros(3)
 
     best_cost = Inf
     best_word = -1
+    best_params = SVector{3,F}(0, 0, 0)
     intermediate_results = DubinsIntermediateResults(q0, q1, ρ)
+    best_path_type = LSL
 
-    path.qi[1] = q0[1]
-    path.qi[2] = q0[2]
-    path.qi[3] = q0[3]
-    path.ρ = ρ
-
-    for i in 0:5
+    for i = 0:5
         path_type = DubinsPathType(i)
-        errcode, params = dubins_word(intermediate_results, path_type)
+        errcode, params = dubins_word(intermediate_results, path_type, loop_tol)
         if errcode == EDUBOK
             cost = sum(params)
             if cost < best_cost
                 best_word = i
                 best_cost = cost
-                path.params[1] = params[1]
-                path.params[2] = params[2]
-                path.params[3] = params[3]
-                path.path_type = path_type
+                best_params = params
+                best_path_type = path_type
             end
         end
     end
 
-    (best_word == -1) && (return EDUBNOPATH, nothing)
+    # exit if no best word
+    (best_word == -1) && (return EDUBNOPATH, DubinsPath{F}())
+
+    # construct the best path
+    path = DubinsPath(q0, best_params, ρ, best_path_type)
+
     return EDUBOK, path
 end
 
@@ -66,32 +66,28 @@ Generate a path with a specified word from an initial configuratioon to a target
 * path_type - the specified path type to use
 * return    - tuple (error code, dubins path). If error code != 0, then `nothing` is returned as the second argument
 """
-function dubins_path(q0::Vector{Float64}, q1::Vector{Float64}, ρ::Float64, path_type::DubinsPathType)
+function dubins_path(
+    q0::VF,
+    q1::VF,
+    ρ::F,
+    path_type::DubinsPathType,
+    loop_tol::F = 1e-5
+) where {F,VF<:AbstractVector{F}}
 
     # input checking
-    @assert length(q0) ==  3
+    @assert length(q0) == 3
     @assert length(q1) == 3
-    (ρ <= 0) && (return EDUBBADRHO, nothing)
+    (ρ <= 0) && (return EDUBBADRHO, DubinsPath{F}())
 
-    path = DubinsPath()
 
     intermediate_results = DubinsIntermediateResults(q0, q1, ρ)
 
-    params = zeros(3)
-    errcode, params = dubins_word(intermediate_results, path_type)
+    errcode, params = dubins_word(intermediate_results, path_type, loop_tol)
     if errcode == EDUBOK
-        path.params[1] = params[1]
-        path.params[2] = params[2]
-        path.params[3] = params[3]
-        path.qi[1] = q0[1]
-        path.qi[2] = q0[2]
-        path.qi[3] = q0[3]
-        path.ρ = ρ
-        path.path_type = path_type
+        path = DubinsPath(q0, params, ρ, path_type)
+        return EDUBOK, path
     end
-
-    (errcode != EDUBOK) && (return errcode, nothing)
-    return errcode, path
+    (errcode != EDUBOK) && (return errcode, DubinsPath{F}())
 end
 
 
@@ -101,7 +97,7 @@ Calculate the length of an initialized path
 * path      - path to find the length of
 * return    - path length
 """
-dubins_path_length(path::DubinsPath) = sum(path.params)*path.ρ
+dubins_path_length(path::DubinsPath) = sum(path.params) * path.ρ
 
 
 """
@@ -111,7 +107,8 @@ Calculate the length of a specific segment of  an initialized path
 * ``i``     - the segment for which the length is required (1-3)
 * return    - segment length
 """
-dubins_segment_length(path::DubinsPath, i::Int) = (i<1 || i>3) ? (return Inf) : (return path.params[i]*path.ρ)
+dubins_segment_length(path::DubinsPath, i::Int) =
+    (i < 1 || i > 3) ? (return Inf) : (return path.params[i] * path.ρ)
 
 """
 Calculate the normalized length of a specific segment of  an initialized path
@@ -120,7 +117,8 @@ Calculate the normalized length of a specific segment of  an initialized path
 * ``i``     - the segment for which the length is required (1-3)
 * return    - normalized segment length
 """
-dubins_segment_length_normalized(path::DubinsPath, i::Int) = (i<1 || i>3) ? (return Inf) : (return path.params[i])
+dubins_segment_length_normalized(path::DubinsPath, i::Int) =
+    (i < 1 || i > 3) ? (return Inf) : (return path.params[i])
 
 """
 Extract the integer that represents which path type was used
@@ -141,30 +139,33 @@ The three operators correspond to ``L``, ``R``, and ``S``
 
  * return    -  the image point as a 3-element vector
 """
-function dubins_segment(t::Float64, qi::Vector{Float64}, segment_type::SegmentType)
+function dubins_segment(
+    t::F,
+    qi::VF,
+    segment_type::SegmentType,
+) where {F,VF<:AbstractVector{F}}
 
-    qt = zeros(3)
     st = sin(qi[3])
     ct = cos(qi[3])
 
     if segment_type == L_SEG
-        qt[1] = +sin(qi[3]+t) - st
-        qt[2] = -cos(qi[3]+t) + ct
-        qt[3] = t
+        qt1 = +sin(qi[3] + t) - st
+        qt2 = -cos(qi[3] + t) + ct
+        qt3 = t
+        qt = SVector{3,F}(qt1, qt2, qt3)
     elseif segment_type == R_SEG
-        qt[1] = -sin(qi[3]-t) + st
-        qt[2] = +cos(qi[3]-t) - ct
-        qt[3] = -t
+        qt1 = -sin(qi[3] - t) + st
+        qt2 = +cos(qi[3] - t) - ct
+        qt3 = -t
+        qt = SVector{3,F}(qt1, qt2, qt3)
     elseif segment_type == S_SEG
-        qt[1] = ct * t
-        qt[2] = st * t
-        qt[3] = 0.0
+        qt1 = ct * t
+        qt2 = st * t
+        qt3 = 0.0
+        qt = SVector{3,F}(qt1, qt2, qt3)
     end
-    qt[1] = qt[1] + qi[1]
-    qt[2] = qt[2] + qi[2]
-    qt[3] = qt[3] + qi[3]
 
-    return qt
+    return qt + qi # should return SVector even if qi is a Vector
 end
 
 """
@@ -174,38 +175,48 @@ Calculate the configuration along the path, using the parameter t
  * ``t``         - length measure where ``0 \\leq t <`` [`dubins_path_length`](@ref)(path)
  * return    - tuple containing non-zero error code if 't' is not in the correct range and the configuration result ``[x, y, \\theta]``
 """
-function dubins_path_sample(path::DubinsPath, t::Float64)
+function dubins_path_sample(path::DubinsPath{F}, t::F; extrapolate = true) where {F}
+
+    # (t < 0 || t > dubins_path_length(path)) && (return EDUBPARAM, path.qi)
+    (t < 0) && (return EDUBPARAM, path.qi)
 
     # tprime is the normalized variant of the parameter t
-    tprime = t/path.ρ
-    qi = zeros(3)
-    q = zeros(3)
-    segment_types = DIRDATA[Int(path.path_type)]
-
-    (t < 0 || t > dubins_path_length(path)) && (return EDUBPARAM, nothing)
+    tprime = t / path.ρ
+    segment_types = DIRDATA[path.path_type]
 
     # initial configuration
-    qi = [0.0, 0.0, path.qi[3]]
+    qi = SVector{3,F}(0, 0, path.qi[3])
 
     # generate target configuration
     p1 = path.params[1]
     p2 = path.params[2]
+    p3 = path.params[3]
     q1 = dubins_segment(p1, qi, segment_types[1])
     q2 = dubins_segment(p2, q1, segment_types[2])
+    q3 = dubins_segment(p3, q2, segment_types[3])
     if tprime < p1
         q = dubins_segment(tprime, qi, segment_types[1])
-    elseif tprime < (p1+p2)
-        q = dubins_segment(tprime-p1, q1, segment_types[2])
+    elseif tprime < (p1 + p2)
+        q = dubins_segment(tprime - p1, q1, segment_types[2])
+    elseif tprime <= (p1 + p2 + p3)
+        q = dubins_segment(tprime - p1 - p2, q2, segment_types[3])
     else
-        q = dubins_segment(tprime-p1-p2, q2, segment_types[3])
+        if extrapolate
+            # extrapolate using straight segment
+            q = dubins_segment(tprime - p1 - p2 - p3, q3, S_SEG)
+        else
+            # not allowed to extrapolate
+            return EDUBPARAM, path.qi
+        end
     end
 
     # scale the target configuration, translate back to the original starting point
-    q[1] = q[1] * path.ρ + path.qi[1]
-    q[2] = q[2] * path.ρ + path.qi[2]
-    q[3] = mod2pi(q[3]);
+    qs1 = q[1] * path.ρ + path.qi[1]
+    qs2 = q[2] * path.ρ + path.qi[2]
+    qs3 = mod2pi(q[3])
+    qs = SVector{3,F}(qs1, qs2, qs3)
 
-    return EDUBOK, q
+    return EDUBOK, qs
 end
 
 
@@ -219,20 +230,20 @@ The sampling process continues until the whole path is sampled, or the callback 
 
  * return       - tuple (error code, configuration vector). If error code != 0, then `nothing` is returned as the second argument
  """
-function dubins_path_sample_many(path::DubinsPath, step_size::Float64)
+function dubins_path_sample_many(path::DubinsPath{F}, step_size::F) where {F}
 
-    q = zeros(3)
-    configurations = []
-    x = 0.0
-    length = dubins_path_length(path)
+    configurations = SVector{3,F}[]
 
-    (step_size < 0 || step_size > length) && (return EDUBPARAM, nothing)
+    L = dubins_path_length(path)
+    (step_size < 0 || step_size > L) && (return EDUBPARAM, configurations)
+
+    x = zero(F)
 
 
-    while x < length
+    while x < L
         errcode, q = dubins_path_sample(path, x)
         push!(configurations, q)
-        (errcode != 0) && (return errcode, nothing)
+        (errcode != 0) && (return errcode, configurations)
         x += step_size
     end
 
@@ -245,7 +256,8 @@ Convenience function to identify the endpoint of a path
  * path          - an initialized path
  * return        - tuple containing (zero on successful completion and the end configuration ``[x,y,\\theta]``)
 """
-dubins_path_endpoint(path::DubinsPath) = dubins_path_sample(path, dubins_path_length(path) - TOL)
+dubins_path_endpoint(path::DubinsPath{F}) where {F} =
+    dubins_path_sample(path, dubins_path_length(path))
 
 """
 Convenience function to extract a sub-path
@@ -254,26 +266,22 @@ Convenience function to extract a sub-path
  * ``t``             - a length measure, where ``0 < t <`` [`dubins_path_length`](@ref)(path)
  * return        - zero on successful completion and the subpath
 """
-function dubins_extract_subpath(path::DubinsPath, t::Float64)
+function dubins_extract_subpath(path::DubinsPath{F}, t::F) where {F}
+
+
+    ((t < 0) || (t > dubins_path_length(path))) && (return EDUBPARAM, path)
 
     # calculate the true parameter
-    tprime = t / path.ρ;
-
-    ((t < 0) || (t > dubins_path_length(path))) && (return EDUBPARAM, nothing)
-
-    newpath = DubinsPath()
-
-    # copy most of the data
-    newpath.qi[1] = path.qi[1]
-    newpath.qi[2] = path.qi[2]
-    newpath.qi[3] = path.qi[3]
-    newpath.ρ = path.ρ
-    newpath.path_type = path.path_type
+    tprime = t / path.ρ
 
     # fix the parameters
-    newpath.params[1] = min(path.params[1], tprime)
-    newpath.params[2] = min(path.params[2], tprime - newpath.params[1])
-    newpath.params[3] = min(path.params[3], tprime - newpath.params[1] - newpath.params[2])
+    newpath_params1 = min(path.params[1], tprime)
+    newpath_params2 = min(path.params[2], tprime - newpath_params1)
+    newpath_params3 = min(path.params[3], tprime - newpath_params1 - newpath_params2)
+    newpath_params = SVector{3,F}(newpath_params1, newpath_params2, newpath_params3)
+
+    # construct the new path (mostly copy old params)
+    newpath = DubinsPath(path.qi, newpath_params, path.ρ, path.path_type)
 
     return EDUBOK, newpath
 end
@@ -284,34 +292,35 @@ The function to call the corresponding Dubins path based on the path_type
 
 * return        - tuple (error code, path length as a vector for corresponding path type)
 """
-function dubins_word(intermediate_results::DubinsIntermediateResults, path_type::DubinsPathType)
+function dubins_word(
+    intermediate_results::DubinsIntermediateResults{F},
+    path_type::DubinsPathType,
+    loop_tol::F
+) where {F}
 
-    result::Int = 0
-    out = Vector{Float64}(undef,3)
     if path_type == LSL
-        result, out = dubins_LSL(intermediate_results)
+        result, out = dubins_LSL(intermediate_results, loop_tol)
     elseif path_type == RSL
-        result, out = dubins_RSL(intermediate_results)
+        result, out = dubins_RSL(intermediate_results, loop_tol)
     elseif path_type == LSR
-        result, out = dubins_LSR(intermediate_results)
+        result, out = dubins_LSR(intermediate_results, loop_tol)
     elseif path_type == RSR
-        result, out = dubins_RSR(intermediate_results)
+        result, out = dubins_RSR(intermediate_results, loop_tol)
     elseif path_type == LRL
-        result, out = dubins_LRL(intermediate_results)
+        result, out = dubins_LRL(intermediate_results, loop_tol)
     elseif path_type == RLR
-        result, out = dubins_RLR(intermediate_results)
+        result, out = dubins_RLR(intermediate_results, loop_tol)
     else
-        result = EDUBNOPATH
+        result, out = EDUBNOPATH, SVector{3,F}(0, 0, 0)
     end
 
-    (result == EDUBNOPATH) && (return result, nothing)
     return result, out
 end
 
-"""
-Reset tolerance value
-"""
-function set_tolerance(ϵ::Float64)
-    TOL = ϵ
-    return
-end
+# """
+# Reset tolerance value
+# """
+# function set_tolerance(ϵ::Float64)
+#     TOL = ϵ
+#     return
+# end