+++ test/runtests.jl
@@ -1,12 +1,13 @@
 using Dubins
-using Memento
-
+using StaticArrays
 using Test
 
-setlevel!(getlogger(Dubins), "error")
+silence!()
 
 
 @testset "Dubins" begin
     include("test_api.jl")
     include("test_path.jl")
+    include("test_allocations.jl")
+    include("test_mod2pi.jl")
 end